divide notes from element_commands
